入力形式予測パイプラインを共通化
