+++ atcodertools/fmtprediction/multi_case_layout.py
@@ -1,71 +1,19 @@
 import re
-from typing import Dict, List, Optional, Tuple
+from typing import List, Optional, Tuple
 
 from bs4 import BeautifulSoup
 
 from atcodertools.client.models.problem_content import (
     ProblemContent,
 )
-from atcodertools.fmtprediction.models.calculator import (
-    EvaluateError,
-)
-from atcodertools.fmtprediction.models.format import Format
-from atcodertools.fmtprediction.models.type import Type
-from atcodertools.fmtprediction.predict_simple_format import (
-    SimpleFormatPredictionFailedError,
-    predict_simple_format,
-)
-from atcodertools.fmtprediction.predict_types import (
-    InvalidLoopIndexError,
-    InvalidLoopSizeError,
-    TooLessFetchesError,
-    TooManyFetchesError,
-    TypePredictor,
-    merge_type_dicts,
-)
-from atcodertools.fmtprediction.token_manager import (
-    TokenManager,
-)
-from atcodertools.fmtprediction.tokenize_format import (
-    NoFormatFoundError,
-    search_formats_with_minimum_vars,
-)
 
-
-MAX_SAMPLE_CASE_COUNT = 100000
 
 _SIMPLE_NAME_RE = re.compile(
     r"^[A-Za-z][A-Za-z0-9_]*$"
 )
-
 _INDEXED_NAME_RE = re.compile(
     r"^([A-Za-z][A-Za-z0-9]*)_([A-Za-z0-9]+)$"
 )
-
-
-class NoMultiCaseFormatFoundError(Exception):
-    pass
-
-
-class MultipleMultiCaseFormatsError(Exception):
-    def __init__(self, candidates):
-        self.candidates = candidates
-
-
-class MultiCaseFormatPrediction:
-    def __init__(
-        self,
-        prefix_format: Format,
-        case_format: Format,
-        case_count_var: str,
-        var_to_type: Dict[str, Type],
-        layout: str,
-    ):
-        self.prefix_format = prefix_format
-        self.case_format = case_format
-        self.case_count_var = case_count_var
-        self.var_to_type = var_to_type
-        self.layout = layout
 
 
 def _nonempty_lines(text: str) -> List[str]:
@@ -327,42 +275,6 @@ def _extract_single_block_indexed_layout(
     return prefix_text, case_text
 
 
-def _predict_simple_format_candidates(
-    input_format_text: str,
-) -> List[Format]:
-    try:
-        tokenized_candidates = (
-            search_formats_with_minimum_vars(
-                input_format_text
-            )
-        )
-    except NoFormatFoundError:
-        return []
-
-    result = []
-    seen = set()
-
-    for tokenized_candidate in tokenized_candidates:
-        for to_1d_flag in (False, True):
-            try:
-                candidate = predict_simple_format(
-                    tokenized_candidate.var_tokens,
-                    to_1d_flag,
-                )
-            except SimpleFormatPredictionFailedError:
-                continue
-
-            signature = str(candidate)
-
-            if signature not in seen:
-                seen.add(signature)
-                result.append(candidate)
-
-            break
-
-    return result
-
-
 def _multi_case_evidence_context(
     content: ProblemContent,
 ) -> str:
@@ -521,200 +433,3 @@ def _candidate_layouts(
                 case_text,
                 False,
             )
-
-
-def _scalar_variable_names(
-    format_: Format,
-) -> List[str]:
-    return [
-        variable.name
-        for variable in format_.all_vars()
-        if variable.dim_num() == 0
-    ]
-
-
-def _format_variable_names(format_: Format):
-    return {
-        variable.name
-        for variable in format_.all_vars()
-    }
-
-
-def _validate_candidate_on_samples(
-    prefix_format: Format,
-    case_format: Format,
-    case_count_var: str,
-    samples,
-):
-    merged_types = {}
-
-    for sample in samples:
-        token_manager = TokenManager(
-            sample.get_input().split()
-        )
-
-        prefix_predictor = TypePredictor(
-            prefix_format
-        )
-        prefix_predictor.consume(
-            token_manager
-        )
-
-        case_count = (
-            prefix_predictor.get_actual_value(
-                case_count_var
-            )
-        )
-
-        if type(case_count) is not int:
-            raise InvalidLoopSizeError
-
-        if (
-            case_count <= 0
-            or case_count > MAX_SAMPLE_CASE_COUNT
-        ):
-            raise InvalidLoopSizeError
-
-        merged_types = merge_type_dicts(
-            merged_types,
-            prefix_predictor.get_typing_result(),
-        )
-
-        for _ in range(case_count):
-            case_predictor = TypePredictor(
-                case_format
-            )
-            case_predictor.consume(
-                token_manager
-            )
-
-            merged_types = merge_type_dicts(
-                merged_types,
-                case_predictor.get_typing_result(),
-            )
-
-        if not token_manager.is_terminal():
-            raise TooManyFetchesError
-
-    return merged_types
-
-
-def predict_multi_case_format(
-    content: ProblemContent,
-) -> MultiCaseFormatPrediction:
-    samples = content.get_samples()
-
-    if not samples:
-        raise NoMultiCaseFormatFoundError
-
-    evidence_context = (
-        _multi_case_evidence_context(
-            content
-        )
-    )
-
-    valid_predictions = []
-    seen = set()
-
-    for (
-        layout,
-        prefix_text,
-        case_text,
-        structural_evidence,
-    ) in _candidate_layouts(content):
-        prefix_candidates = (
-            _predict_simple_format_candidates(
-                prefix_text
-            )
-        )
-        case_candidates = (
-            _predict_simple_format_candidates(
-                case_text
-            )
-        )
-
-        for prefix_format in prefix_candidates:
-            prefix_names = (
-                _format_variable_names(
-                    prefix_format
-                )
-            )
-
-            for case_count_var in (
-                _scalar_variable_names(
-                    prefix_format
-                )
-            ):
-                if (
-                    not structural_evidence
-                    and not _context_supports_count_variable(
-                        evidence_context,
-                        case_count_var,
-                    )
-                ):
-                    continue
-
-                for case_format in case_candidates:
-                    case_names = (
-                        _format_variable_names(
-                            case_format
-                        )
-                    )
-
-                    if prefix_names.intersection(
-                        case_names
-                    ):
-                        continue
-
-                    try:
-                        var_to_type = (
-                            _validate_candidate_on_samples(
-                                prefix_format,
-                                case_format,
-                                case_count_var,
-                                samples,
-                            )
-                        )
-                    except (
-                        AssertionError,
-                        EvaluateError,
-                        InvalidLoopIndexError,
-                        InvalidLoopSizeError,
-                        KeyError,
-                        NotImplementedError,
-                        StopIteration,
-                        TooLessFetchesError,
-                        TooManyFetchesError,
-                        ValueError,
-                    ):
-                        continue
-
-                    signature = (
-                        str(prefix_format),
-                        str(case_format),
-                        case_count_var,
-                    )
-
-                    if signature in seen:
-                        continue
-
-                    seen.add(signature)
-                    valid_predictions.append(
-                        MultiCaseFormatPrediction(
-                            prefix_format,
-                            case_format,
-                            case_count_var,
-                            var_to_type,
-                            layout,
-                        )
-                    )
-
-    if not valid_predictions:
-        raise NoMultiCaseFormatFoundError
-
-    if len(valid_predictions) > 1:
-        raise MultipleMultiCaseFormatsError(
-            valid_predictions
-        )
-
-    return valid_predictions[0]
